dune/requirements.txt
google-auth
google-auth-oauthlib
requests
rich
groq
textual
